Research Assistant/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.6.1
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.5.1
openai==1.12.0
python-dotenv==1.0.1
whoosh==2.7.4
requests==2.31.0
numpy==1.26.4
pandas==2.2.0
streamlit==1.31.1
beautifulsoup4==4.12.3
transformers==4.37.2
torch==2.2.0 